firewall: fix the three things Phase 3 could only learn from the compiler
7d85eae47 landed the merged engine written against signatures derived from the
vendored Go. Three were wrong, and the Kotlin compiler named all three exactly.

1. Bridge was missing onDNSAdded / onDNSRemoved / onDNSStopped. DNSListener
   EMBEDS ResolverListener (backend/dnsx.go), and gomobile flattens embedded
   interfaces onto the same Java type. My extraction matched method signatures
   — lines with a capital letter and a paren — so a bare `ResolverListener`
   line was invisible to it. Checked the other four union members the same way
   afterwards: none of them embed anything, so the union is 21 methods, not 18.

2. Intra.connect3(fd, mtu, bridge) takes two Longs, and the staged call passed
   Ints. The mapping rule was right — Go `int` becomes Kotlin `Long` — I had
   just applied it to the interfaces I wrote and not to the call site I
   inherited. Now .toLong() on both.

3. FirewallController.cloudVpn did not exist. FirestackTunnelService and
   FirewallFlowBridge both read it, and CloudVpn.kt defines the interface, but
   nothing ever declared the property. Added as a @Volatile @JvmStatic var
   defaulting to CloudVpnProvider.NONE, not lateinit: under NONE every
   "vpn-only" app is simply blocked, which is the safe reading when no tunnel
   is configured. A firewall that throws on an unset provider fails open.

The pattern in all three is the same: everything derivable from the Go source I
derived correctly, and every miss was somewhere the source does not show the
shape — an embedded interface, an inherited call site, a property no file
declares. That is what the compile step is for, and why the previous commit
said plainly that none of it had been compiled.

Guard suite PASS.

+++ ab_cloud-libs-shared/libs/firewall/src/main/java/com/diegonmarcos/superapp/firewall/FirestackTunnelService.kt
@@ -64,7 +64,7 @@ class FirestackTunnelService : VpnService() {
         tun = fd
         tunnel = runCatching {
             // Connect3(fd, tunmtu, bridge): the simplest firestack entry.
-            Intra.connect3(fd.fd, TUN_MTU, buildBridge())
+            Intra.connect3(fd.fd.toLong(), TUN_MTU.toLong(), buildBridge())
         }.onFailure { Log.e(TAG, "firestack connect failed", it) }.getOrNull()
 
         if (tunnel == null) { FirewallPrefs.setEnabled(ctx, false); teardown(); stopSelf(); return }
